markdownify/pyproject.toml
[project]
name = "markdownify-mcp"
version = "0.1.0"
description = "Streamable HTTP MCP server for converting CSV/XLSX/PDF to Markdown with optional images"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "mcp[http]>=0.2.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pandas>=2.2.3",
  "tabulate>=0.9.0",
  "openpyxl>=3.1.5",
  "pdfplumber>=0.11.4",
  "pymupdf>=1.24.11",
  "pillow>=11.0.0",
  "python-multipart>=0.0.9",
]

[project.scripts]
markdownify-gateway = "gateway:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["markdownify_app"]
include = ["gateway.py"]

[tool.ruff]
target-version = "py313"
line-length = 120
extend-exclude = ["sessions"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E203"]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
  "pytest>=8.3.0",
]
